requirements.txt
numpy
python-telegram-bot
matplotlib